+++ core/src/com/group12/cookiesrising/gameworld/GameWorld.java
@@ -199,14 +199,14 @@ public class GameWorld {
                 currentEnemy.action(party);
                 for(Hero h: party.getHeroList()){
                     if (h.isAlive()){
-                        dmgTextPool.getDamageText(Double.toString(Math.floor(currentEnemy.getAttackPoint())),Math.round(h.getPosition().x),Math.round(h.getPosition().y));
+                        dmgTextPool.getDamageText(Integer.toString((int)Math.floor(currentEnemy.getAttackPoint())),Math.round(h.getPosition().x),Math.round(h.getPosition().y));
                     }
                 }
             }else {
                 Hero h = party.getHeroList().get(target);
                 currentEnemy.action(h);
                 if (h.isAlive()){
-                    dmgTextPool.getDamageText(Double.toString(Math.floor(currentEnemy.getAttackPoint())),Math.round(h.getPosition().x),Math.round(h.getPosition().y));
+                    dmgTextPool.getDamageText(Integer.toString((int)Math.floor(currentEnemy.getAttackPoint())),Math.round(h.getPosition().x),Math.round(h.getPosition().y));
                 }
             }
         }
